% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as_psych}
\alias{as_psych}
\alias{as_psych.sfa}
\title{Coerce to psych fa Object}
\usage{
as_psych(x, ...)

\method{as_psych}{sfa}(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{c("psych", "fa")}.
}
\description{
Coerce to psych fa Object
}
