% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{sfa_naming_instruction}
\alias{sfa_naming_instruction}
\title{The Default Naming Instruction}
\usage{
sfa_naming_instruction()
}
\value{
A character scalar.
}
\description{
Returns the instruction string used to embed items for factor naming.
}
